+++ pycinema/filters/ImageReader.py
@@ -63,7 +63,6 @@ class ImageReader(Filter):
                         raise ValueError('h5 file not formatted correctly')
                     for k in group.keys():
                         data = numpy.array(group.get(k))
-                        # print('xxx',data)
                         if data.dtype == '|S10' and len(data)==1:
                             data = data[0].decode('UTF-8')
                         # elif len(data)==1:
